removed incomplete tutorial, fixed float casting

+++ metal/end_model/loss.py
@@ -27,13 +27,13 @@ class SoftCrossEntropyLoss(nn.Module):
     def forward(self, input, target):
         n, k = input.shape
         # Note that t.new_zeros, t.new_full put tensor on same device as t
-        cum_losses = input.new_zeros(n).float()
+        cum_losses = input.new_zeros(n)
         for y in range(k):
             cls_idx = input.new_full((n,), y, dtype=torch.long)
             y_loss = F.cross_entropy(input, cls_idx, reduction="none")
             if self.weight is not None:
                 y_loss = y_loss * self.weight[y]
-            cum_losses += target[:, y].float() * y_loss.float()
+            cum_losses += target[:, y].float() * y_loss
         if self.reduction == "none":
             return cum_losses
         elif self.reduction == "elementwise_mean":
